chore: Update bookmark collections

+++ application/database/migrations/2025_06_16_091811_update_collection_tabel.php
@@ -1,8 +1,9 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
 return new class extends Migration
 {
@@ -12,11 +13,25 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('bookmark_collections', function (Blueprint $table) {
-            $table->dropColumn('parent_id');
-            $table->foreignId('fund_id')->nullable()->constrained('funds');
-            $table->renameColumn('type_id',to: 'model_id');
-            $table->renameColumn('type', to: 'model_type');
-            $table->renameColumn('type_type', to: 'type');
+            if (Schema::hasColumn('bookmark_collections', 'parent_id')) {
+                $table->dropColumn('parent_id');
+            }
+
+            if (!Schema::hasColumn('bookmark_collections', 'fund_id')) {
+                $table->foreignId('fund_id')->nullable()->constrained('funds');
+            }
+
+            if (Schema::hasColumn('bookmark_collections', 'type_id') && !Schema::hasColumn('bookmark_collections', 'model_id')) {
+                DB::statement("ALTER TABLE bookmark_collections RENAME COLUMN type_id TO model_id");
+            }
+
+            if (Schema::hasColumn('bookmark_collections', 'type') && !Schema::hasColumn('bookmark_collections', 'model_type')) {
+                DB::statement("ALTER TABLE bookmark_collections RENAME COLUMN type TO model_type");
+            }
+
+            if (Schema::hasColumn('bookmark_collections', 'type_type') && !Schema::hasColumn('bookmark_collections', 'type')) {
+                DB::statement("ALTER TABLE bookmark_collections RENAME COLUMN type_type TO type");
+            }
         });
     }
 
@@ -26,6 +41,7 @@ return new class extends Migration
     public function down(): void
     {
         // Schema::table('bookmarks_collections', function (Blueprint $table) {
+        //     //
         // });
     }
 };
